Write meta to ROFI_DATA instead of ROFI_INFO Rofi provides two ways to pass meta data from one script call to the next: - Defining a `data` field in the header of the script output, which is   accessible as environment variable `ROFI_DATA` during the next run of   the script. - Defining a `info` field in a menu entry line of the script output,   which is accessible as `ROFI_INFO` during the next run of the script.
Until now, rofi used the second method to save the meta data, which
leads to the problem, that all metadata are lost, if a custom menu
entry is chosen, since those are passed to the next script run as the
raw text, typed into the text field, without any prepared entry line, in
which hidden infos could be stored.

So the easiest thing seems to be, to move all non-item-dependent values
to the header, to keep the script functional, if custom items are
involved.